requirements.txt
asgiref==3.9.1
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
crispy-bootstrap4==2025.6
cryptography==45.0.7
cssselect2==0.8.0
Django==5.2.6
django-allauth==65.11.1
django-countries==7.6.1
django-crispy-forms==2.4
django-environ==0.12.0
et_xmlfile==2.0.0
fonttools==4.59.2
gunicorn==23.0.0
idna==3.10
msgpack==1.1.1
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
razorpay==1.4.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
